Minor bug fixes, added pause functionality to audio sub system. Collisions now save last collision normal in the collider.
Resolved few minor warnings.

Finished school project crunching. Starting class hierarchy revision.

+++ Rusty SDL/RustyEngine/ColliderCircle.cpp
@@ -47,7 +47,7 @@ namespace RustyEngine
 	void ColliderCircle::setRadius(float r)
 	{
 		radius = r;
-		area = M_PI * pow(radius, 2);
+		area = (float)M_PI * pow(radius, 2);
 	}
